refactored `data` into new `active_learning` script
